pyproject.toml
[tool.poetry]
name = "f1guru"
version = "0.1.0"
description = ""
authors = ["Pau Matas <pmatasalbiol@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.13"
mysql-connector-python = "^8.3.0"
langchain-openai = "^0.1.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
